+++ Tests/YamsTests/RepresenterTests.swift
@@ -98,6 +98,13 @@ class RepresenterTests: XCTestCase {
         XCTAssertEqual(Node("test"), "test")
     }
 
+    func testUUID() throws {
+        try XCTAssertEqual(
+            Node(UUID(uuidString: "B5C6C790-BC0A-4781-9AFF-F9896E0C030C")),
+            "B5C6C790-BC0A-4781-9AFF-F9896E0C030C"
+        )
+    }
+
     func testOptional() throws {
         XCTAssertEqual(try Node(Int?.none), "null")
     }
@@ -126,6 +133,7 @@ extension RepresenterTests {
             ("testFloat", testFloat),
             ("testInteger", testInteger),
             ("testString", testString),
+            ("testUUID", testUUID),
             ("testOptional", testOptional),
             ("testArray", testArray),
             ("testDictionary", testDictionary)