+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Swift-3-branch
+## 3.0.1
 
 ##### Breaking